chore: move fs cheatcodes to own file
